# open-api/tabliodb-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments/threads": {
      "post": {
        "operationId": "createCommentThread",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentThreadCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/diagram/{diagramId}/summary": {
      "get": {
        "operationId": "getCommentDiagramSummary",
        "parameters": [
          {
            "name": "diagramId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDiagramSummaryDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/diagram/{diagramId}/threads": {
      "get": {
        "operationId": "getCommentThreads",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "diagramId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/threads/{threadId}/comments": {
      "get": {
        "operationId": "getThreadComments",
        "parameters": [
          {
            "name": "parentCommentId",
            "required": false,
            "in": "query",
            "description": "Use null to list root comments, a comment id to list direct replies, or omit it for legacy flat lists.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "replyToCommentThread",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentReplyCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/threads/{threadId}/root-comments": {
      "get": {
        "operationId": "getCommentThreadRootComments",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/comments/{commentId}/replies": {
      "get": {
        "operationId": "getCommentReplies",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "replyToComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentReplyCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/comments/{commentId}": {
      "patch": {
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/threads/{threadId}/read-state": {
      "get": {
        "operationId": "getCommentThreadReadState",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadReadStateDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/threads/{threadId}/read": {
      "patch": {
        "operationId": "markCommentThreadRead",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadReadStateDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/threads/{threadId}/resolve": {
      "patch": {
        "operationId": "resolveCommentThread",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadStatusResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/threads/{threadId}/unresolve": {
      "patch": {
        "operationId": "unresolveCommentThread",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentThreadStatusResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tabliodb API",
    "description": "Collaborative database schema designer API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tabliodb_access_token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "CommentLexicalDocumentDto": {
        "type": "object",
        "properties": {
          "root": {
            "type": "object",
            "properties": {
              "children": {
                "maxItems": 200,
                "type": "array",
                "items": {}
              },
              "type": {
                "type": "string",
                "enum": [
                  "root"
                ]
              }
            },
            "required": [
              "children",
              "type"
            ],
            "additionalProperties": {}
          }
        },
        "required": [
          "root"
        ],
        "additionalProperties": {}
      },
      "CommentThreadCreateDto": {
        "type": "object",
        "properties": {
          "bodyJson": {
            "$ref": "#/components/schemas/CommentLexicalDocumentDto"
          },
          "diagramId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "targetType": {
            "type": "string",
            "enum": [
              "diagram",
              "table",
              "column",
              "relationship",
              "index",
              "enum",
              "check",
              "note",
              "group"
            ]
          }
        },
        "required": [
          "bodyJson",
          "diagramId",
          "targetId",
          "targetType"
        ]
      },
...
      "CommentThreadResponseDto_Output": {
        "type": "object",
        "properties": {
          "thread": {
            "type": "object",
            "properties": {
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "createdById": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "diagramId": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "resolvedAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "nullable": true
              },
              "resolvedById": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "nullable": true
              },
              "status": {
                "type": "string",
                "enum": [
                  "open",
                  "resolved"
                ]
              },
              "targetId": {
                "type": "string",
                "nullable": true
              },
              "targetType": {
                "type": "string",
                "enum": [
                  "diagram",
                  "table",
                  "column",
                  "relationship",
                  "index",
                  "enum",
                  "check",
                  "note",
                  "group"
                ]
              },
              "unreadCount": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              }
            },
            "required": [
              "createdAt",
              "createdById",
              "diagramId",
              "id",
              "resolvedAt",
              "resolvedById",
              "status",
              "targetId",
              "targetType",
              "unreadCount",
              "updatedAt"
            ],
            "additionalProperties": false
          },
          "comment": {
            "$ref": "#/components/schemas/CommentResponseDto_Output"
          }
        },
        "required": [
          "thread",
          "comment"
        ],
        "additionalProperties": false
      },
      "CommentThreadTargetSummaryDto_Output": {
        "type": "object",
        "properties": {
          "openCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "resolvedCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "targetType": {
            "type": "string",
            "enum": [
              "diagram",
              "table",
              "column",
              "relationship",
              "index",
              "enum",
              "check",
              "note",
              "group"
            ]
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "unreadCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "nullable": true
          }
        },
        "required": [
          "openCount",
          "resolvedCount",
          "targetId",
          "targetType",
          "totalCount",
          "unreadCount",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "CommentDiagramSummaryDto_Output": {
        "type": "object",
        "properties": {
          "diagramId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "openCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "resolvedCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentThreadTargetSummaryDto_Output"
            }
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "unreadCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "nullable": true
          }
        },
        "required": [
          "diagramId",
          "openCount",
          "resolvedCount",
          "targets",
          "totalCount",
          "unreadCount",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "CommentThreadListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentThreadListItemDto_Output"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "additionalProperties": false
      },
      "CommentListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponseDto_Output"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "additionalProperties": false
      },
      "CommentReplyCreateDto": {
        "type": "object",
        "properties": {
          "bodyJson": {
            "$ref": "#/components/schemas/CommentLexicalDocumentDto"
          },
          "parentCommentId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "nullable": true
          }
        },
        "required": [
          "bodyJson"
        ]
      },
      "CommentUpdateDto": {
        "type": "object",
        "properties": {
          "bodyJson": {
            "$ref": "#/components/schemas/CommentLexicalDocumentDto"
          }
        },
        "required": [
          "bodyJson"
        ]
      },
...
    }
  }
}